+++ leaspy/models/multivariate_parallel_model.py
@@ -30,7 +30,7 @@ class MultivariateParallelModel(AbstractMultivariateModel):
         for k in parameters.keys():
             if k in ['mixing_matrix']:
                 continue
-            self.parameters[k] = torch.tensor(parameters[k], dtype=torch.float32)
+            self.parameters[k] = torch.tensor(parameters[k])
         self.attributes = LogisticParallelAttributes(self.name, self.dimension, self.source_dimension)
         self.attributes.update(['all'], self.parameters)
